Move all render item classes to their own namespace and folder.
