+++ Commands/DeleteCommand.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using MetaphysicsIndustries.Ligra.RenderItems;
 
 namespace MetaphysicsIndustries.Ligra.Commands
 {
