+++ Commands/FuncAssignCommand.cs
@@ -1,4 +1,5 @@
 using System.Linq;
+using MetaphysicsIndustries.Ligra.RenderItems;
 using MetaphysicsIndustries.Solus;
 
 namespace MetaphysicsIndustries.Ligra.Commands
