+++ Commands/TSolveCommand.cs
@@ -1,3 +1,4 @@
+using MetaphysicsIndustries.Ligra.RenderItems;
 using MetaphysicsIndustries.Solus;
 
 namespace MetaphysicsIndustries.Ligra.Commands
