+++ RenderItems/ControlItem.cs
@@ -1,11 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Drawing;
-
-using MetaphysicsIndustries.Solus;
-
-namespace MetaphysicsIndustries.Ligra
+namespace MetaphysicsIndustries.Ligra.RenderItems
 {
     //public class ControlItem : RenderItem
     //{
