+++ RenderItems/Graph3dItem.cs
@@ -1,10 +1,9 @@
 using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Drawing;
 using MetaphysicsIndustries.Solus;
 
-namespace MetaphysicsIndustries.Ligra
+namespace MetaphysicsIndustries.Ligra.RenderItems
 {
     public class Graph3dItem : RenderItem
     {
