+++ RenderItems/GraphMatrixItem.cs
@@ -1,12 +1,10 @@
 using System;
 using System.Collections.Generic;
-using System.Text;
-using MetaphysicsIndustries.Solus;
 using System.Drawing;
-
 using MetaphysicsIndustries.Acuity;
+using MetaphysicsIndustries.Solus;
 
-namespace MetaphysicsIndustries.Ligra
+namespace MetaphysicsIndustries.Ligra.RenderItems
 {
     public class GraphMatrixItem : RenderItem
     {
