+++ RenderItems/GraphVectorItem.cs
@@ -1,13 +1,10 @@
 using System;
 using System.Collections.Generic;
-using System.Text;
-using MetaphysicsIndustries.Solus;
-//using MetaphysicsIndustries.Utilities;
 using System.Drawing;
-
 using MetaphysicsIndustries.Acuity;
+using MetaphysicsIndustries.Solus;
 
-namespace MetaphysicsIndustries.Ligra
+namespace MetaphysicsIndustries.Ligra.RenderItems
 {
     public class GraphVectorItem : RenderItem
     {
